structure updated, refactoring
