+++ examples/ddd-config-emtry-dirs-example.yaml
