+++ examples/ddd-config-minimal-example.yaml
