+++ examples/ddd-config-simple-example.yaml
