+++ examples/ddd-config-standard-example.yaml
