+++ src/templates/config_templates/ddd-config-advanced.yaml
@@ -1,0 +1,58 @@
+settings:
+  preset: "advanced"
+  contexts_layout: "nested"
+  init_imports: true
+  group_components: false
+  root_name: "src"
+
+layers:
+  contexts:
+    - name: user_management
+      domain:
+        entities:
+        value_objects:
+        services:
+        repositories:
+        events:
+        exceptions:
+
+      application:
+        use_cases:
+        event_handlers:
+        validators:
+        exceptions:
+
+      infrastructure:
+        repositories:
+        models:
+        message_bus:
+        background_tasks:
+        configs:
+
+      interface:
+        controllers:
+        dto:
+        api_routes:
+        middleware:
+        error_handlers:
+
+  # Shared Kernel - Common components across contexts
+  shared_kernel:
+    domain:
+      value_objects:
+      events:
+      exceptions:
+      interfaces: I
+
+    application:
+      interfaces:
+      exceptions:
+
+    infrastructure:
+      caching:
+      logging:
+
+    interface:
+      responses:
+      filters:
+      middleware:
