+++ src/templates/config_templates/ddd-config-simple.yaml
@@ -1,0 +1,43 @@
+settings:
+  preset: "simple"
+  init_imports: true
+
+layers:
+  # Domain Layer - Core business logic
+  domain:
+    entities:
+    value_objects:
+    services:
+    repositories:
+    events:
+    exceptions:
+
+  # Application Layer - Use cases and business workflows
+  application:
+    use_cases:
+    event_handlers:
+    validators:
+    exceptions:
+
+  # Infrastructure Layer - Technical implementations
+  infrastructure:
+    repositories:
+    models:
+    message_bus:
+    background_tasks:
+    configs:
+
+  # Interface Layer - External facing components
+  interface:
+    controllers:
+    dto:
+    api_routes:
+    middleware:
+    api_error_handlers:
+
+  # Common/Shared components
+  common:
+    constants:
+    exceptions:
+    utils:
+    interfaces:
