+++ src/templates/config_templates/ddd-config-standard.yaml
@@ -1,0 +1,55 @@
+settings:
+  preset: "standard"
+  init_imports: true
+  group_components: true
+
+layers:
+  # Domain Layer - Organized by bounded contexts
+  domain:
+    contexts:
+      - name: user_management
+        entities:
+        value_objects:
+        services:
+        repositories:
+        events:
+        exceptions:
+
+
+
+  # Application Layer - Use cases grouped by context
+  application:
+    contexts:
+      - name: user_management
+        use_cases:
+        event_handlers:
+        validators:
+        exceptions:
+
+
+  # Infrastructure Layer - Technical implementations
+  infrastructure:
+    repositories:
+    models:
+    message_bus:
+    background_tasks:
+    configs:
+
+  # Interface Layer - API and external interfaces
+  interface:
+    contexts:
+      - name: user_management
+        controllers:
+        dto:
+        api_routes:
+
+    # Shared interface components
+    middleware:
+    error_handlers:
+
+  # Common/Shared components
+  common:
+    constants:
+    exceptions:
+    utils:
+    interfaces:
